Adding some functions & fixing

+++ srcs/ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: varichar <varichar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/06 14:54:56 by varichar          #+#    #+#             */
-/*   Updated: 2016/08/06 16:48:15 by varichar         ###   ########.fr       */
+/*   Updated: 2016/08/06 17:24:28 by varichar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
