+++ srcs/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: varichar <varichar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/06 15:57:04 by varichar          #+#    #+#             */
-/*   Updated: 2016/08/06 16:49:25 by varichar         ###   ########.fr       */
+/*   Updated: 2016/08/06 17:25:13 by varichar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
