+++ srcs/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: varichar <varichar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/04 13:14:44 by varichar          #+#    #+#             */
-/*   Updated: 2016/08/06 16:49:03 by varichar         ###   ########.fr       */
+/*   Updated: 2016/08/06 17:26:23 by varichar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int		i;
+	size_t		i;
 	void				*pt;
 
 	pt = s;
